Tests under /rs/nervous_system compile
TODO:
1. Figure out if we'll get conflicts again if we introduce a
canisters--test_feature target; the SNS swap is the biggest potential cause of problems
2. Decide whether we want canisters as a subdir of common
3. Update Cargo.toml files to add the dependency on canisters
