+++ rs/nervous_system/canisters/BUILD.bazel
@@ -1,0 +1,59 @@
+load("@rules_rust//rust:defs.bzl", "rust_library")
+
+package(default_visibility = ["//visibility:public"])
+
+DEPENDENCIES = [
+    # Keep sorted.
+    "//packages/icrc-ledger-types:icrc_ledger_types",
+    "//rs/nervous_system/common",
+    "//rs/nervous_system/runtime",
+    "//rs/nns/constants",
+    "//rs/rosetta-api/icp_ledger",
+    "//rs/rosetta-api/ledger_core",
+    "//rs/rust_canisters/dfn_core",
+    "//rs/rust_canisters/dfn_protobuf",
+    "//rs/types/base_types",
+    # "//rs/crypto/sha2",
+    #"//rs/rust_canisters/http_types",
+    # "@crate_index//:base64",
+    # "@crate_index//:build-info",
+    # "@crate_index//:by_address",
+    # "@crate_index//:bytes",
+    # "@crate_index//:ic-metrics-encoder",
+    # "@crate_index//:ic-stable-structures",
+    # "@crate_index//:json5",
+    # "@crate_index//:lazy_static",
+    # "@crate_index//:maplit",
+    # "@crate_index//:mockall",
+    # "@crate_index//:num-traits",
+    # "@crate_index//:priority-queue",
+    # "@crate_index//:prost",
+    # "@crate_index//:rust_decimal",
+    # "@crate_index//:serde",
+    # "@crate_index//:serde_json",
+]
+
+MACRO_DEPENDENCIES = [
+    # Keep sorted.
+    "@crate_index//:async-trait",
+    "@crate_index//:rust_decimal_macros",
+]
+
+rust_library(
+    name = "canisters",
+    srcs = glob(["src/**"]),
+    crate_name = "ic_nervous_system_canisters",
+    proc_macro_deps = MACRO_DEPENDENCIES,
+    version = "0.9.0",
+    deps = DEPENDENCIES,
+)
+
+rust_library(
+    name = "canisters--test_feature",
+    srcs = glob(["src/**"]),
+    crate_features = ["test"],
+    crate_name = "ic_nervous_system_canisters",
+    proc_macro_deps = MACRO_DEPENDENCIES,
+    version = "0.9.0",
+    deps = DEPENDENCIES,
+)
