+++ rs/nervous_system/canisters/src/ledger.rs
@@ -1,0 +1,147 @@
+use async_trait::async_trait;
+use dfn_core::{call, CanisterId};
+use dfn_protobuf::protobuf;
+use ic_ledger_core::block::BlockIndex;
+use ic_nervous_system_common::ledger::{ICRC1Ledger, IcpLedger};
+use ic_nervous_system_common::NervousSystemError;
+use icp_ledger::{
+    tokens_from_proto, AccountBalanceArgs, AccountIdentifier, Memo, SendArgs,
+    Subaccount as IcpSubaccount, Tokens, TotalSupplyArgs,
+};
+use icrc_ledger_types::icrc1::account::{Account, Subaccount};
+
+pub struct IcpLedgerCanister {
+    id: CanisterId,
+}
+
+impl IcpLedgerCanister {
+    pub fn new(id: CanisterId) -> Self {
+        IcpLedgerCanister { id }
+    }
+}
+
+#[async_trait]
+impl ICRC1Ledger for IcpLedgerCanister {
+    async fn transfer_funds(
+        &self,
+        amount_e8s: u64,
+        fee_e8s: u64,
+        from_subaccount: Option<Subaccount>,
+        to: Account,
+        memo: u64,
+    ) -> Result<BlockIndex, NervousSystemError> {
+        <IcpLedgerCanister as IcpLedger>::transfer_funds(
+            self,
+            amount_e8s,
+            fee_e8s,
+            from_subaccount.map(IcpSubaccount),
+            icrc1_account_to_icp_accountidentifier(to),
+            memo,
+        )
+        .await
+    }
+
+    async fn total_supply(&self) -> Result<Tokens, NervousSystemError> {
+        <IcpLedgerCanister as IcpLedger>::total_supply(self).await
+    }
+
+    async fn account_balance(&self, account: Account) -> Result<Tokens, NervousSystemError> {
+        <IcpLedgerCanister as IcpLedger>::account_balance(
+            self,
+            icrc1_account_to_icp_accountidentifier(account),
+        )
+        .await
+    }
+
+    fn canister_id(&self) -> CanisterId {
+        self.id
+    }
+}
+
+#[async_trait]
+impl IcpLedger for IcpLedgerCanister {
+    async fn transfer_funds(
+        &self,
+        amount_e8s: u64,
+        fee_e8s: u64,
+        from_subaccount: Option<IcpSubaccount>,
+        to: AccountIdentifier,
+        memo: u64,
+    ) -> Result<u64, NervousSystemError> {
+        // Send 'amount_e8s' to the target account.
+        //
+        // We expect the 'fee_e8s' AND 'amount_e8s' to be
+        // deducted from the from_subaccount. When calling
+        // this method, make sure that the staked amount
+        // can cover BOTH of these amounts, otherwise there
+        // will be an error.
+        let result: Result<u64, (Option<i32>, String)> = call(
+            self.id,
+            "send_pb",
+            protobuf,
+            SendArgs {
+                memo: Memo(memo),
+                amount: Tokens::from_e8s(amount_e8s),
+                fee: Tokens::from_e8s(fee_e8s),
+                from_subaccount,
+                to,
+                created_at_time: None,
+            },
+        )
+        .await;
+
+        result.map_err(|(code, msg)| {
+            NervousSystemError::new_with_message(format!(
+                "Error calling method 'send' of the ledger canister. Code: {:?}. Message: {}",
+                code, msg
+            ))
+        })
+    }
+
+    async fn total_supply(&self) -> Result<Tokens, NervousSystemError> {
+        let result: Result<Tokens, (Option<i32>, String)> =
+            call(self.id, "total_supply_pb", protobuf, TotalSupplyArgs {})
+                .await
+                .map(tokens_from_proto);
+
+        result.map_err(|(code, msg)| {
+            NervousSystemError::new_with_message(
+                format!(
+                    "Error calling method 'total_supply' of the ledger canister. Code: {:?}. Message: {}",
+                    code, msg
+                )
+            )
+        })
+    }
+
+    async fn account_balance(
+        &self,
+        account: AccountIdentifier,
+    ) -> Result<Tokens, NervousSystemError> {
+        let result: Result<Tokens, (Option<i32>, String)> = call(
+            self.id,
+            "account_balance_pb",
+            protobuf,
+            AccountBalanceArgs { account },
+        )
+        .await
+        .map(tokens_from_proto);
+
+        result.map_err(|(code, msg)| {
+            NervousSystemError::new_with_message(
+                format!(
+                    "Error calling method 'account_balance_pb' of the ledger canister. Code: {:?}. Message: {}",
+                    code, msg
+                )
+            )
+        })
+    }
+
+    fn canister_id(&self) -> CanisterId {
+        self.id
+    }
+}
+
+fn icrc1_account_to_icp_accountidentifier(account: Account) -> AccountIdentifier {
+    AccountIdentifier::new(account.owner.into(), account.subaccount.map(IcpSubaccount))
+}
