+++ rs/nervous_system/canisters/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod cmc;
+pub mod ledger;
