+++ rs/nns/governance/BUILD.bazel
@@ -216,6 +216,7 @@ rust_canister(
     deps = DEPENDENCIES + [
         ":build_script",
         ":governance",
+        "//rs/nervous_system/canisters",
     ],
 )
