+++ rs/rust_canisters/dfn_core/BUILD.bazel
@@ -5,6 +5,7 @@ load("//bazel:defs.bzl", "rust_ic_test")
 package(default_visibility = [
     # Keep sorted.
     "//publish/canisters:__pkg__",
+    "//rs/nervous_system/canisters:__subpackages__",
     "//rs/nervous_system/clients:__pkg__",
     "//rs/nervous_system/common:__subpackages__",
     "//rs/nervous_system/root:__pkg__",
