+++ rs/sns/swap/canister/canister.rs
@@ -7,6 +7,7 @@ use dfn_core::{
 use ic_base_types::PrincipalId;
 use ic_canister_log::log;
 use ic_canisters_http_types::{HttpRequest, HttpResponse, HttpResponseBuilder};
+use ic_nervous_system_canisters::ledger::IcpLedgerCanister;
 use ic_nervous_system_clients::{
     canister_id_record::CanisterIdRecord,
     canister_status::CanisterStatusResultV2,
@@ -354,8 +355,8 @@ fn now_seconds() -> u64 {
 /// Returns a real ledger stub that communicates with the specified
 /// canister, which is assumed to be the ICP production ledger or a
 /// canister that implements that same interface.
-fn create_real_icp_ledger(id: CanisterId) -> ic_nervous_system_common::ledger::IcpLedgerCanister {
-    ic_nervous_system_common::ledger::IcpLedgerCanister::new(id)
+fn create_real_icp_ledger(id: CanisterId) -> IcpLedgerCanister {
+    IcpLedgerCanister::new(id)
 }
 
 #[export_name = "canister_init"]
